Remove depedency on pkg_resources and import __version__ as needed.
pkg_resources comes from setuptools, but is only used in one place, when
setting the user agent in the public key download.  Instead, we can just
import the __version__ in the __init__.py module file as needed, for example:

from scitokens import __version__ as PKG_VERSION

Be careful of circular dependencies.  It is advisable to import this within the
function that you need the PKG_VERSION only.